requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Werkzeug==3.0.3
openai
python-dateutil==2.9.0
python-dotenv==1.0.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
